Separate operator tests
